# Refs #28459 -- Improved performance of time difference expressions on MySQL.