@adamchainz 2018-03-05
Simple and effective, nice!